requirements.txt
chatterbot==0.1.1
hackernews-python==0.3.2
